Moved Tree.cs to Contracts
