+++ WebApi/Triggers/TreeTriggers.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using ForestGEO.WebApi.Model.Storage;
+using ForestGEO.WebApi.Model.Contracts;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
@@ -35,7 +36,7 @@ namespace ForestGEO.WebApi.Triggers.Tree
             log.LogInformation("C# HTTP trigger function processed a request.");
 
             string requestBody = await new StreamReader(req.Body).ReadToEndAsync();
-            var data = JsonConvert.DeserializeObject<Tree[]>(requestBody);
+            var data = JsonConvert.DeserializeObject<ForestGEO.WebApi.Model.Contracts.Tree[]>(requestBody);
 
             Dictionary<(string,string),ForestGEO.WebApi.Model.Storage.TreeStorage> TreeRecords = mySql.QueryTreeDB("select TempID, QuadratName, Tag, StemTag, Mnemonic as SpCode, DBH, Codes, HOM, ExactDate, x, y, PlotID, CensusID, Errors from tempnewplants");
